feat: update validate api value

+++ backend/src/main/java/com/ndt/AcademicAdvising/services/impl/PostServiceImpl.java
@@ -4,13 +4,14 @@
  */
 package com.ndt.AcademicAdvising.services.impl;
 
-import com.ndt.AcademicAdvising.dto.RequestPostDTO;
 import com.ndt.AcademicAdvising.dto.ResponsePostDTO;
+import com.ndt.AcademicAdvising.dto.ResponseUserDTO;
 import com.ndt.AcademicAdvising.pojo.Post;
+import com.ndt.AcademicAdvising.pojo.User;
 import com.ndt.AcademicAdvising.repositories.PostRepository;
 import com.ndt.AcademicAdvising.repositories.UserRepository;
 import com.ndt.AcademicAdvising.services.PostService;
-import java.util.Optional;
+import java.util.Objects;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
@@ -24,20 +25,30 @@ import org.springframework.stereotype.Service;
  * @author ngodo
  */
 @Service
-public class PostServiceImpl implements PostService{
-    
+public class PostServiceImpl implements PostService {
+
     @Autowired
     private PostRepository postRepo;
-    
+
     @Autowired
     private UserRepository userRepo;
-    
-    public ResponsePostDTO toDTO(Post post) {
+
+    private ResponsePostDTO toDTO(Post post) {
         ResponsePostDTO dto = new ResponsePostDTO();
         dto.setId(post.getId());
         dto.setContent(post.getContent());
         dto.setCreatedAt(post.getCreatedAt());
+        dto.setUser(new ResponseUserDTO(post.getUser().getName(), 
+                                        post.getUser().getUsername(), 
+                                        post.getUser().getEmail(), 
+                                        post.getUser().getAvatar(), 
+                                        post.getUser().getStudentCode()));
         return dto;
+    }
+
+    private User getCurrentUser() {
+        String username = SecurityContextHolder.getContext().getAuthentication().getName();
+        return this.userRepo.findByUsername(username);
     }
 
     @Override
@@ -48,24 +59,58 @@ public class PostServiceImpl implements PostService{
 
     @Override
     public ResponsePostDTO addPost(String content) {
-        Post post = new Post();
-        post.setContent(content);
-//        post.setUser(this.userRepo.findByUsername(SecurityContextHolder.getContext().getAuthentication().getName()));
-        post.setUser(this.userRepo.findByUsername("ngodothanh"));
-        this.postRepo.save(post);
-        return toDTO(post);
+        if (content == null || content.isBlank()) {
+            throw new IllegalArgumentException("Nội dung không được để trống.");
+        } else {
+            Post post = new Post();
+            post.setContent(content);
+            post.setUser(getCurrentUser());
+            this.postRepo.save(post);
+            return toDTO(post);
+        }
     }
-    
+
     @Override
-    public Boolean existPost(int postId) {
-        return this.postRepo.existsById(postId);
+    public ResponsePostDTO updatePost(int postId, String content) {
+        if (content == null || content.isBlank()) {
+            throw new IllegalArgumentException("Nội dung không được để trống.");
+        }
+
+        Post post = postRepo.findById(postId)
+                .orElseThrow(() -> new IllegalArgumentException("Không tìm thấy bài viết."));
+
+        User currentUser = getCurrentUser();
+
+        if (!post.getUser().getId().equals(currentUser.getId())) {
+            throw new IllegalArgumentException("Không thể cập nhật bài đăng của người khác.");
+        }
+        
+        post.setContent(content);
+        postRepo.save(post);
+        return toDTO(post);
     }
 
     @Override
     public void deletePost(int postId) {
-        this.postRepo.deleteById(postId);
+        Post p = this.postRepo.findById(postId)
+                .orElseThrow(() -> new IllegalArgumentException("Bài đăng không tồn tại."));
+        
+        User currentUser = getCurrentUser();
+        
+        if (Objects.equals(p.getUser().getId(), currentUser.getId())) {
+            this.postRepo.deleteById(postId);
+        } else {
+            throw new IllegalArgumentException("Không thể xoá bài đăng của người khác.");
+        }
     }
 
-
+    @Override
+    public Page<ResponsePostDTO> getListPostUser() {
+        User currentUser = getCurrentUser();
+        Pageable pageable = PageRequest.of(0, 10);
+        
+        return this.postRepo.findAllByUserIdOrderByCreatedAtDesc(pageable, currentUser.getId())
+                .map(this::toDTO);
+    }
 
 }
